+++ src/expr-parser/src/parser.rs
@@ -1384,7 +1384,7 @@ mod analyses {
         let lookahead = input.lookahead1();
 
         let scalar_type = if input.look_and_eat(bigint, &lookahead) {
-            ReprScalarType::Int64
+            ReprScalarType::Int
         } else if input.look_and_eat(double, &lookahead) {
             input.parse::<precision>()?;
             ReprScalarType::Float64
@@ -1394,9 +1394,9 @@ mod analyses {
             input.parse::<varying>()?;
             ReprScalarType::String
         } else if input.look_and_eat(integer, &lookahead) {
-            ReprScalarType::Int32
+            ReprScalarType::Int
         } else if input.look_and_eat(smallint, &lookahead) {
-            ReprScalarType::Int16
+            ReprScalarType::Int
         } else if input.look_and_eat(text, &lookahead) {
             ReprScalarType::String
         } else {
